requirements_txt.txt
Flask==2.3.3
line-bot-sdk==3.5.0
google-generativeai==0.3.2
requests==2.31.0
gunicorn==21.2.0